Add functionality to create and add new tasks to the inbox section
This commit adds the necessary code to create and add new tasks to the inbox. The code imports and uses the PageCreator, DomController, TaskLibrary, and TaskForm modules. It creates the necessary objects and appends the task form to the inbox page content. The code also includes event listeners to show and hide the add task form and add new tasks to the library. The createTask method is used to create a new task object, which is then added to the TaskLibrary using the addTask method. Finally, the inbox tasks are displayed using the showTasks method from the DomController module.