Add functionality to add and display projects
This commit adds new functionality that allows users to create and display new projects. A new 'add project' button functionality has been added, which opens a form when clicked. When the user enters a new project name and submits the form, the project is added to the task library, and the list of projects is updated and displayed. Users can now select a project by clicking on the corresponding element in the project list, which updates the category and project selectors accordingly.